refactor: Move SquaredTrajectory to its own file
